% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableWorkspaces.R
\name{availableWorkspaces}
\alias{availableWorkspaces}
\title{Available workspaces}
\usage{
availableWorkspaces(
  accountEmail = gcloud_account(),
  billingProjectName = gcloud_project()
)
}
\value{
A tibble with available workspaces.
}
\description{
Available workspaces
}
