% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseByCategory.R
\name{browseByCategory}
\alias{browseByCategory}
\title{Browsing workspaces based on values under the selected category}
\usage{
browseByCategory(target, category, display_categories = FALSE)
}
\arguments{
\item{target}{A character(1) specifying the AnVIL resource you want to
perform searching. Available options are \code{workspaces}, \code{workflows},
and \code{data}.}

\item{category}{A metadata category you want to browse.}

\item{display_categories}{A logical. Default is \code{FALSE}. If it
is set to \code{TRUE}, the function returns the character vector
of available \code{category} option for a selected \code{target}.}
}
\description{
Browsing workspaces based on values under the selected category
}
\examples{
browseByCategory(target = "workspaces", category = "accessLevel")

}
