% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseByKeyword.R
\name{browseByKeyword}
\alias{browseByKeyword}
\title{Browsing keyword in AnVIL workspaces, workflows, and data}
\usage{
browseByKeyword(keyword, updateAllMeta = FALSE)
}
\arguments{
\item{keyword}{}
}
\description{
Browsing keyword in AnVIL workspaces, workflows, and data
}
\examples{
browseByCategory(keyword = "CNV")

}
