% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findInputName.R
\name{.findInputName}
\alias{.findInputName}
\title{Find the root entity name}
\usage{
.findInputName(workspaceName, rootEntity, nameOnly = TRUE)
}
\arguments{
\item{workspaceName}{Name of the workspace}

\item{rootEntity}{A character. Type of root entity for Terra's data model.
For example, \code{participant}, \code{participant_set}, \code{sample}, etc.}

\item{nameOnly}{Under the default (\code{TRUE}), only the names of a given
root entity type will be returned.}
}
\description{
Find the root entity name
}
