% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_workflow_fullname}
\alias{.get_workflow_fullname}
\title{Get the workflow namespace and name}
\usage{
.get_workflow_fullname(workspaceName, workflowName = NULL)
}
\arguments{
\item{workspaceName}{A character. Name of the workspace to use.}

\item{workflowName}{A character. Name of the workflow to run. If a single
workflow is available under the selected workspace, this function will
check the input of that workflow under the default (\code{NULL}). If there
are multiple workflows available, you should specify the workflow.}
}
\value{
A character of \code{workflow_namespace/workflow_name}
}
\description{
Use this internally when \code{\link{setCloudEnv}} is already run.
}
\keyword{internal}
