% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_workspace_fullname}
\alias{.get_workspace_fullname}
\title{Get the fullname of the workspace}
\usage{
.get_workspace_fullname(workspaceName)
}
\arguments{
\item{workspaceName}{Character(1). Name of the template workspace name you
want to clone. You can provide \code{name} or \code{namespace/name}.}
}
\value{
Character(1) of \code{workspaceNamespace/workspaceName}
}
\description{
Get the fullname of the workspace
}
\examples{
library(AnVILBase)
if (has_avworkspace(strict = TRUE, platform = AnVILGCP::gcp())) {
.get_workspace_fullname(workspaceName = "Bioconductor-Workflow-DESeq2")
}

}
\keyword{internal}
