% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaTables.R
\name{.getAgeMinMax}
\alias{.getAgeMinMax}
\title{Apply to the returned value from \code{.avSampleData} function}
\usage{
.getAgeMinMax(sampleData)
}
\arguments{
\item{sampleData}{A data frame.}
}
\description{
Apply to the returned value from \code{.avSampleData} function
}
