% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getMethodConfig}
\alias{.getMethodConfig}
\title{Get method configuration}
\usage{
.getMethodConfig(billingProjectName, workspaceName)
}
\arguments{
\item{workspaceName}{A character (1). Name of the workspace you want to
get method configuration.}

\item{workspaceNamespace}{A character (1). Namespace of the workspace you
want to get method configuration.}
}
\value{
A data frame containing method configuration. \code{name} and \code{namespace}
slots contain workspacename and workspacenamespace, respectively.
}
\description{
Get method configuration
}
