% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.mostRecentSucceededSubmissionId}
\alias{.mostRecentSucceededSubmissionId}
\title{Get the submission id of the most recently succeeded submission}
\usage{
.mostRecentSucceededSubmissionId(billingProjectName, workspaceName)
}
\arguments{
\item{billingProjectName}{Name of the billing project}

\item{workspaceName}{Name of the workspace you want to get method configuration.}
}
\value{
Submission Id of the most recent submission
}
\description{
Get the submission id of the most recently succeeded submission
}
