% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnVILBrowse.R
\name{.search_keyword}
\alias{.search_keyword}
\title{Search keywords in a given metadata table}
\usage{
.search_keyword(keyword, metadata)
}
\arguments{
\item{keyword}{A character(1). Regular expression is accepted. For example,
you can search multiple keywords separated by the vertical bar ("\code{|}").}

\item{metadata}{A data frame. Metadata table of workspace, workflow, or
AnVIL data.}
}
\value{
A data frame. A subset of input metadata table with the rows
containing the keyword.
}
\description{
Search keywords in a given metadata table
}
