% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop_quietly}
\alias{.stop_quietly}
\title{Stop the execution without error messages}
\usage{
.stop_quietly()
}
\value{
Stop the function call without warning/error messages.
}
\description{
Stop the execution without error messages
}
\keyword{internal}
