% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaTables.R
\name{.tableToString}
\alias{.tableToString}
\title{Summary table into a single string where the column name and the value
is separated by \code{sep} and columns are separated by \code{delim}}
\usage{
.tableToString(tb, sep = ":", delim = ";")
}
\arguments{
\item{tb}{A table. Output from the \code{table} function.}

\item{sep}{A delimiter to separate the column name and value}

\item{delim}{A delimiter to separate columns}
}
\description{
Summary table into a single string where the column name and the value
is separated by \code{sep} and columns are separated by \code{delim}
}
