% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectAllMeta.R
\name{getAllDataTables}
\alias{getAllDataTables}
\title{Get all the data tables}
\usage{
getAllDataTables(workspaces = NULL)
}
\arguments{
\item{workspaces}{A character vector. Under the default (\code{NULL}),
all the data tables from all the workspaces user has access to will be
returned. If you specify this, the data tables only from the specified
workspace(s) will be returned.}
}
\value{
A Data Frame of all the data tables
}
\description{
Get all the data tables
}
\examples{
library(AnVILBase)
if (
    GCPtools::gcloud_exists() && identical(avplatform_namespace(), "AnVILGCP") &&
    nzchar(avworkspace_name())
) {
allDataTables <- getAllDataTables()
}

}
