% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectAllMeta.R
\name{getAllWorkflows}
\alias{getAllWorkflows}
\title{Collect workflows from all workspaces a user has access to}
\usage{
getAllWorkflows(workspaces = NULL)
}
\arguments{
\item{workspaces}{Under the default (\code{NULL}), workflows from all the
workspaces a user has access to will be collected.}
}
\description{
Collect workflows from all workspaces a user has access to
}
\examples{
library(AnVILBase)
if (has_avworkspace(strict = TRUE, platform = AnVILGCP::gcp())) {
allWorkflows <- getAllWorkflows()
}

}
