% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaTables.R
\name{getData}
\alias{getData}
\title{Creates a metadata table of data from all workspaces provided}
\usage{
getData(allWorkspaces)
}
\arguments{
\item{allWorkspaces}{A data frame of all the workspaces you have access
to. An output from the \code{getWorkspaces} function.}
}
\description{
This function usually takes a long time to run due to the large volume
of AnVIL data.
}
