% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataTables.R
\name{getDataTables}
\alias{getDataTables}
\title{Check the current input arguments}
\usage{
getDataTables(workspaceName, table = NULL)
}
\arguments{
\item{workspaceName}{Character(1). Name of the workspace that contains
the workflow(s) you want to launch.}

\item{table}{Character(1). Any value under the \code{table} column of the
output with the default \code{table = NULL} can be used.}
}
\value{
A data frame. It contains all the data table(s) associated with
the workspace/table.
}
\description{
Check the current input arguments
}
\examples{
library(AnVIL)
if (gcloud_exists() && nzchar(avworkspace_name())) {
getDataTables(workspaceName = "Bioconductor-Workflow-DESeq2")
}

}
