% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaTables.R
\name{getWorkflows}
\alias{getWorkflows}
\title{Creates a metadata table of workflows from all workspaces provided}
\usage{
getWorkflows(allWorkspaces)
}
\arguments{
\item{allWorkspaces}{A data frame of all the workspaces you have access
to. An output from the \code{getWorkspaces} function.}
}
\description{
Creates a metadata table of workflows from all workspaces provided
}
