% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOutput.R
\name{listOutput}
\alias{listOutput}
\title{List outputs}
\usage{
listOutput(
  accountEmail,
  billingProjectName,
  workspaceName,
  submissionId = NULL,
  keyword = NULL,
  includeMetadata = FALSE
)
}
\arguments{
\item{accountEmail}{Email linked to Terra account}

\item{billingProjectName}{Name of the billing project}

\item{workspaceName}{Name of the workspace}

\item{submissionId}{Submission Id. If it's not provided, the most recent
submission id with the 'succeeded' status will be used.}

\item{keyword}{A character string containing a regular expression to be matched
in the output file name. Under the default \code{NULL}, all the outputs from
the workflow, including log files, will be returned.}

\item{includeMetadata}{Under the default (\code{FALSE}), metadata files (e.g.
\code{stderr, stdout, .log, .sh}), will not be returned.}
}
\value{
A tibble with four columns
\itemize{
\item file : character() 'base name' of the file in the bucket.
\item workflow : character() name of the workflow the file is associated with.
\item task : character() name of the task in the workflow that generated the file.
\item path : character() full path to the file in the google bucket.
}
}
\description{
List outputs
}
