% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map}
\alias{map}
\title{The configuration of the workspaces manually curated in this package}
\format{
A data frame containing the details of the workspace configuration
}
\usage{
map
}
\description{
The configuration of the workspaces manually curated in this package
}
\author{
Sehyun Oh \email{shbrief@gmail.com}
}
\keyword{datasets}
