% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorSubmission.R
\name{monitorSubmission}
\alias{monitorSubmission}
\title{Check the status of submitted jobs}
\usage{
monitorSubmission(accountEmail, billingProjectName, workspaceName)
}
\arguments{
\item{accountEmail}{Email linked to Terra account}

\item{billingProjectName}{Name of the billing project}

\item{workspaceName}{Name of the workspace}
}
\value{
A tibble summarizing submitted workflow jobs. Contains information such
as submission Id, submission date, and submission status.
}
\description{
Check the status of submitted jobs
}
