% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCloudEnv.R
\name{setCloudEnv}
\alias{setCloudEnv}
\title{Setup Google Cloud Account and Project}
\usage{
setCloudEnv(
  accountEmail = GCPtools::gcloud_account(),
  billingProjectName = GCPtools::gcloud_project(),
  message = TRUE
)
}
\arguments{
\item{accountEmail}{Character(1). Email linked to your Terra account.}

\item{billingProjectName}{Character(1). Name of the billing project, which
is the gcloud account.}

\item{message}{Under the default (\code{TRUE}), this function will print out
Google Cloud Account and Billing Project set in the working environment}
}
\value{
Terra/AnVIL working environment - Google Cloud billing account
and the billing project name - will be printed out.
}
\description{
Setup Google Cloud Account and Project
}
\examples{
if (GCPtools::gcloud_exists()) {
setCloudEnv()
}

}
