% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnVILBrowse.R
\name{AnVILBrowse}
\alias{AnVILBrowse}
\title{Seach AnVIL workspaces using keywords}
\usage{
AnVILBrowse(
  keyword,
  returnFrom = "all",
  metaTables = "default",
  workspaceTable = NULL,
  workflowTable = NULL,
  dataTable = NULL
)
}
\arguments{
\item{keyword}{A character(1). Regular expression is accepted. For example,
you can search multiple keywords separated by the vertical bar ("\code{|}").}

\item{returnFrom}{Under the default (\code{all}), all the workspaces
containing keywords in their workspace/workflow/data will be returned.
The other available options are \code{workspace}, \code{workflow}, and
\code{data}.}

\item{metaTables}{Under the default (\code{default}), all the publicly
accessible AnVIL workspaces will be subjected for search. If you want
to search in all the workspaces you have access to, set this argument as
\code{custom}, and provide the inputs for \code{workspaceTable},
\code{workflowTable}, and \code{dataTable} arguments.}

\item{workspaceTable}{A data frame. This argument is counted only when
\code{metaTables = "custom"}. Provide the output from the
\code{getWorkspaces} function, to search in all the workspaces
you have access to.}

\item{workflowTable}{A data frame. This argument is counted only when
\code{metaTables = "custom"}. Provide the output from the
\code{getWorkflows} function, to search in all the workflows
you have access to.}

\item{dataTable}{A data frame. This argument is counted only when
\code{metaTables = "custom"}. Provide the output from the \code{getData}
function, to search in all the AnVIL data you have access to.}
}
\value{
A data frame of AnVIL resources containing keywords. Depending on
the \code{returnFrom} argument, it can be workspaces, workflows, or data.
}
\description{
Seach AnVIL workspaces using keywords
}
\examples{
AnVILBrowse("malaria")
AnVILBrowse("resistance")
AnVILBrowse("resistance", returnFrom = "workflows")

}
