% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableAnalysis.R
\name{availableAnalysis}
\alias{availableAnalysis}
\title{Find the available analysis}
\usage{
availableAnalysis(simplify = TRUE)
}
\arguments{
\item{simplify}{Default is \code{TRUE}. If it is set to \code{FALSE}, the
additional information on workspace and workflow will be printed too.}
}
\value{
A data frame. The \code{analysis} columns shows the name of the
available analyses, which is the required input (\code{analysis} argument)
for the functions implemented in AnVILWorkflow package.
}
\description{
This function shows the available analyses and the brief
descriptions of them.
}
\examples{
availableAnalysis()

}
