% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneWorkspace.R
\name{cloneWorkspace}
\alias{cloneWorkspace}
\title{Clone template workspace}
\usage{
cloneWorkspace(
  workspaceName,
  analysis,
  accountEmail = gcloud_account(),
  billingProjectName = gcloud_project()
)
}
\arguments{
\item{workspaceName}{Name of the workspace you are creating}

\item{analysis}{Name of the analysis you want to clone it's workspace. The
list of available analyses can be found using \code{availableAnalysis()}.}

\item{accountEmail}{Email linked to Terra account}

\item{billingProjectName}{Name of the billing project}
}
\value{
Name of the cloned workspace
}
\description{
Clone template workspace
}
