% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneWorkspace.R
\name{cloneWorkspace}
\alias{cloneWorkspace}
\title{Clone template workspace}
\usage{
cloneWorkspace(accountEmail, billingProjectName, workspaceName, analysis)
}
\arguments{
\item{accountEmail}{Email linked to Terra account}

\item{billingProjectName}{Name of the billing project}

\item{workspaceName}{Name of the workspace you are creating}

\item{analysis}{Name of the analysis you want to clone it's workspace. The
list of available analyses can be found using \code{availableAnalysis()}.}
}
\value{
The name of cloned workspace
}
\description{
Clone template workspace
}
