% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentInput.R
\name{currentInput}
\alias{currentInput}
\title{Check the current input arguments}
\usage{
currentInput(
  workspaceName,
  workflowName = NULL,
  accountEmail = gcloud_account(),
  billingProjectName = gcloud_project(),
  requiredInputOnly = TRUE,
  analysis = NULL
)
}
\arguments{
\item{workspaceName}{Name of the workspace}

\item{workflowName}{Name of the workflow to run. If a single workflow is
available under the specified workspace, this function will check the input
of that workflow under the default (\code{NULL}). If there are multiple
workflows available, you should specify the workflow.}

\item{accountEmail}{Email linked to Terra account}

\item{billingProjectName}{Name of the billing project}

\item{requiredInputOnly}{Under the default (\code{TRUE}), only the required
inputs are returned.}

\item{analysis}{If specified, only the minimally required inputs for a
given workflow will be returned.}
}
\value{
A data.frame for the inputs defined in a workflow configuration.
}
\description{
Check the current input arguments
}
