% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentInput.R
\name{.biobakery_currentInput}
\alias{.biobakery_currentInput}
\title{Check the current input arguments}
\usage{
.biobakery_currentInput(config)
}
\arguments{
\item{config}{Workflow configuration. Output from the
\code{\link{getWorkflowConfig}} function.}
}
\value{
A list length of two, including inputListPath and inputFilePath.
}
\description{
Check the current input arguments
}
\examples{
library(AnVILBase)
if (
    gcloud_exists() && identical(avplatform_namespace(), "AnVILGCP") &&
    nzchar(avworkspace_name())
) {
config <- avworkflow_configuration_get(
workflow_namespace = "mtx_workflow_biobakery_version3", 
workflow_name = "mtx_workflow_biobakery_version3", 
workspace_namespace = "waldronlab-terra-rstudio", 
workspace_name = "mtx_workflow_biobakery_version3_template")
biobakery_inputs <- .biobakery_currentInput(config)
}

}
\keyword{internal}
