% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentInput.R
\name{.biobakery_currentInput}
\alias{.biobakery_currentInput}
\title{Check the current input arguments}
\usage{
.biobakery_currentInput(config)
}
\arguments{
\item{config}{Terra workflow configuration. Output from the
\code{avworkflow_configuration_get()} function.}
}
\description{
Check the current input arguments
}
