% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.nonMetadataOutputs}
\alias{.nonMetadataOutputs}
\title{Subset to non-metadata output files}
\usage{
.nonMetadataOutputs(workflowOutputs)
}
\arguments{
\item{workflowOutputs}{A data frame of workflow outputs with four columns - file,
workflow, task, and path. Returned value from \code{\link[AnVIL]{avworkflow_files}}.}
}
\value{
A character vector containing the names of non-metadata output files
}
\description{
Subset to non-metadata output files
}
