% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectAllMeta.R
\name{getAllWorkspaces}
\alias{getAllWorkspaces}
\title{Get AnVIL workspaces}
\usage{
getAllWorkspaces()
}
\description{
Different from \code{\link[AnVIL]{avworkspaces}}https://drive.google.com/drive/u/0/folders/1NNAzcNRBx4nPfcdqjKPeUlVE7lhxXMeL
}
\examples{
library(AnVIL)
if (gcloud_exists() && nzchar(avworkspace_name())) {
allWorkspaces <- getAllWorkspaces()
}

}
