% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectAllMeta.R
\name{getAllWorkspaces}
\alias{getAllWorkspaces}
\title{Get AnVIL workspaces}
\usage{
getAllWorkspaces()
}
\description{
Different from \code{\link[AnVILGCP:avworkspace-methods]{avworkspaces}}
}
\examples{
library(AnVIL)
if (gcloud_exists() && nzchar(avworkspace_name())) {
allWorkspaces <- getAllWorkspaces()
}

}
