% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDashboard.R
\name{getDashboard}
\alias{getDashboard}
\title{Print out Dashboard contents}
\usage{
getDashboard(analysis)
}
\arguments{
\item{analysis}{The name of the analysis use want to check the Dashboard of.
The list of available analyses can be found with \code{availableAnalysis()}.}
}
\value{
The last modified date as a message, followed by the Dashboard
contents of the target workspace.
}
\description{
This function prints out the Dashboard contents of the target workspace.
}
\examples{
getDashboard("salmon")

}
