% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutput.R
\name{getOutput}
\alias{getOutput}
\title{Download output files from Terra}
\usage{
getOutput(
  workspaceName,
  submissionId = NULL,
  keyword = NULL,
  includeMetadata = FALSE,
  dest_dir = ".",
  dry = TRUE
)
}
\arguments{
\item{workspaceName}{Name of the workspace}

\item{submissionId}{Submission Id. If it's not provided, the most recent
submission id with the 'succeeded' status will be used.}

\item{keyword}{A character string containing a regular expression to be matched
in the output file name. Under the default \code{NULL}, all the outputs from
the workflow, including log files, will be returned.}

\item{includeMetadata}{Under the default (\code{FALSE}), metadata files (e.g.
\code{stderr, stdout, .log, .sh}), will not be returned.}

\item{dest_dir}{Path to the directory where downloaded files are saved}

\item{dry}{To download the output data, set \code{dry = FALSE}.}
}
\description{
Download output files from Terra
}
