% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorWorkflow.R
\name{monitorWorkflow}
\alias{monitorWorkflow}
\title{Check the status of submitted jobs}
\usage{
monitorWorkflow(workspaceName)
}
\arguments{
\item{workspaceName}{Name of the workspace}
}
\value{
A tibble summarizing submitted workflow jobs. Contains information
such as submission Id, submission date, and submission status.
}
\description{
Check the status of submitted jobs
}
