% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWorkflow.R
\name{runWorkflow}
\alias{runWorkflow}
\title{Launch Terra workflow}
\usage{
runWorkflow(
  workspaceName,
  workflowName = NULL,
  useCallCache = TRUE,
  inputName = NULL,
  accountEmail = gcloud_account(),
  billingProjectName = gcloud_project()
)
}
\arguments{
\item{workspaceName}{Name of the workspace that contains the workflow(s) you
want to launch.}

\item{workflowName}{Name of the workflow to run.}

\item{useCallCache}{A logical. Under the default condition (\code{TRUE}), call
cache will be used.}

\item{inputName}{Name of you input entity. If the workflow is using Terra's
data model, this is required. The available entities can be found using the
\code{findInputName} function.}

\item{accountEmail}{Email linked to Terra account}

\item{billingProjectName}{Name of the billing project}
}
\description{
Launch Terra workflow
}
