% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableHead.R
\name{tableHead}
\alias{tableHead}
\title{Display the head of output tsv file}
\usage{
tableHead(
  x,
  n = 6,
  workspaceName,
  submissionId = NULL,
  accountEmail = gcloud_account(),
  billingProjectName = gcloud_project()
)
}
\arguments{
\item{x}{Output file name}

\item{n}{The number of rows to return. Default is 6.}

\item{workspaceName}{Name of the workspace}

\item{submissionId}{Submission Id. If it's not provided, the most recent
submission id will be used.}

\item{accountEmail}{Email linked to Terra account}

\item{billingProjectName}{Name of the billing project}
}
\description{
If outputs contain \verb{\code{.tsv}} files, you can check the head of those files
using this function without downloading them.
}
