% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInput.R
\name{updateInput}
\alias{updateInput}
\title{Update the input}
\usage{
updateInput(
  workspaceName,
  inputs,
  workflowName = NULL,
  dry = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{workspaceName}{Name of the workspace}

\item{inputs}{A tibble containing new input values. Provide the modify
version of the current input table, which is a returned value from
\code{\link{currentInput}} function.}

\item{workflowName}{Name of the workflow to run. If a single workflow is
available under the specified workspace, this function will check the input
of that workflow under the default (\code{NULL}). If there are multiple
workflows available, you should specify the workflow.}

\item{dry}{Logical(1). When \code{TRUE} (default), report the updated
configuration but do not perform the action requested in Terra. When
\code{FALSE}, inputs in Terra/AnVIL will updated.}

\item{verbose}{Logical(1). When \code{TRUE} (default), this function
will print the updated input.}
}
\value{
With \code{verbose=TRUE}, a list of updated inputs will be
printed. A successful execution of the function will update the input
configuration of the target workflow in Terra/AnVIL.
}
\description{
Update the input
}
\examples{
library(AnVIL)
if (gcloud_exists() && nzchar(avworkspace_name())) {
if ("salmon" \%in\% avworkspaces()$name) {
inputs <- currentInput(workspaceName = "salmon")
## Modify the contents of 'inputs' table for your analysis
updateInput(workspaceName = "salmon", inputs = inputs) 
}}

}
